Commented EVERYTHING.

+++ commands/owner/say.js
@@ -18,9 +18,11 @@ module.exports = class SayCommand extends Command {
 		});
 	}
 
-	async run(msg,args) {
+	async run(msg,args) { //Echos the written content.
 		const { text } = args;
-		msg.delete();
-		return msg.say(`\u180E${text}`);
+		if(msg.author.id == `147604925612818432`){
+			msg.delete();
+			return msg.say(`\u180E${text}`);
+		} else {msg.say(`Excuse me who do you think you are`);}
 	}
 };
